fix(catalog): cast smallint+1 back to smallint to avoid INT4 promotion
Postgres promotes `smallint + integer-literal` to INT4, so the prior
SELECT returned INT4 even though the inner term was explicitly
`::smallint`.  sqlx then rejected the decode with `Rust type 'i16'
(as SQL type 'INT2') is not compatible with SQL type 'INT4'` and the
register handler 500'd.  Wrap the whole expression in `::smallint` so
the column type matches the `i16` binding.

Surfaced during noetl/ai-meta#54 Phase F R5 e2e re-probe — every
POST to /api/catalog/register failed against the Rust server in the
local kind cluster.

Closes noetl/server#68
Refs noetl/ai-meta#54

+++ src/db/queries/catalog.rs
@@ -10,9 +10,14 @@ use crate::error::AppResult;
 /// matches the column type and avoids the sqlx decode mismatch that
 /// surfaced during noetl/ai-meta#49 Phase A ui_schema validation.
 pub async fn get_next_version(pool: &DbPool, path: &str) -> AppResult<i16> {
+    // `smallint + integer-literal` returns `INT4` in Postgres, so we cast
+    // the entire expression back to `smallint` to match the `i16` binding
+    // the sqlx decoder expects.  Without the outer cast sqlx errors with
+    // `Rust type 'i16' (as SQL type 'INT2') is not compatible with SQL
+    // type 'INT4'`.
     let result: Option<(i16,)> = sqlx::query_as(
         r#"
-        SELECT COALESCE(MAX(version), 0)::smallint + 1
+        SELECT (COALESCE(MAX(version), 0)::smallint + 1)::smallint
         FROM noetl.catalog
         WHERE path = $1
         "#,
